Move code to path utils.
